將驗證部分進行大幅度更改

+++ routes/web.php
@@ -26,16 +26,18 @@ Route::get('/forest', [ForestController::class,'index']);
 
 Route::get('/forest/{id}', [ForestController::class, 'show'])->where('id', '[0-9]+')->name('Forest.show');
     // 修改表單
-Route::get('/forest/{id}/edit', [ForestController::class, 'edit'])->where('id', '[0-9]+')->name('Forest.edit');
-Route::delete('/forest/delete/{id}', [ForestController::class, 'destroy'])->where('id', '[0-9]+')->name('Forest.destroy');
+
+Route::delete('/forest/delete/{id}', [ForestController::class, 'destroy'])->where('id', '[0-9]+')->name('Forest.destroy')->middleware('can:admin');
 
 Route::patch('forest/update/{id}', [ForestController::class, 'update'])->name('forest.update');
 
-Route::get('/forest/create', [ForestController::class, 'create'])->name('Forest.create')->middleware('can:admin');
+Route::get('/forest/create', [ForestController::class, 'create'])->name('Forest.create')->middleware('can:admin' );
 
 Route::post('forest/store', [ForestController::class, 'store'])->name('Forest.store');
 
-Route::get('forest/{id}/edit', [ForestController::class, 'edit'])->where('id', '[0-9]+')->name('Forest.edit')->middleware('can:manger');
+Route::get('forest/{id}/edit', [ForestController::class, 'edit'])->where('id', '[0-9]+')->name('Forest.edit')->middleware('can:edit-forest');
+
+
 Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
